requirements.txt
pygame-ce>=2.1.3
chess>=1.9
stockfish>=3
Pillow>=9.0
